[Improvement] Move DorisSourceBuilder into DorisSource as a static inner builder class
